--- src/com/ua/robot/GIT/Main.java
@@ -1,12 +1,0 @@
-package com.ua.robot.GIT;
-
-public class Main {
-    public static void main(String[] args) {
-        int x = 21;
-        int y = 15;
-        System.out.println(x + y);
-        System.out.println(x - y);
-        System.out.println(x / y);
-        System.out.println(x * y);
-    }
-}
