+++ src/com/ua/robot/Main.java
@@ -1,0 +1,18 @@
+package com.ua.robot;
+
+public class Main {
+    public static void main(String[] args) {
+        int x = 21;
+        int y = 15;
+        System.out.println(x + y);
+        System.out.println(x - y);
+        System.out.println(x / y);
+        System.out.println(x * y);
+        System.out.println("_____________________________");
+        String str = "imagination";
+        System.out.println(str.toUpperCase());
+        System.out.println(str.toLowerCase());
+        System.out.println(str.charAt(0));
+        System.out.println(str.charAt(str.length() - 1));
+    }
+}
